Calculate texture byte sizes as 64-bit on GL and Vulkan. Closes #3338

+++ renderdoc/driver/vulkan/vk_pixelhistory.cpp
@@ -1243,7 +1243,7 @@ protected:
     else
     {
       viewInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-      uint32_t bs = GetByteSize(1, 1, 1, format, 0);
+      uint32_t bs = (uint32_t)GetByteSize(1, 1, 1, format, 0);
 
       if(bs == 1)
         viewInfo.format = VK_FORMAT_R8_UINT;
